hop-node: add stateUpdater to all watchers and simplify
