+++ packages/hop-node/src/watchers/StakeWatcher.ts
@@ -125,8 +125,9 @@ class StakeWatcher extends BaseWatcher {
             )} canonical token to L2 hop token`
           )
 
-          if (this.dryMode) {
-            this.logger.warn('dry mode: skipping approve transaction')
+          await this.handleStateSwitch()
+          if (this.isDryOrPauseMode) {
+            this.logger.warn(`dry: ${this.dryMode}, pause: ${this.pauseMode}. skipping approve`)
             return
           }
 
@@ -195,8 +196,9 @@ class StakeWatcher extends BaseWatcher {
       )
     }
     this.logger.debug(`attempting to stake ${formattedAmount} tokens`)
-    if (this.dryMode) {
-      this.logger.warn('dry mode: skipping stake transaction')
+    await this.handleStateSwitch()
+    if (this.isDryOrPauseMode) {
+      this.logger.warn(`dry: ${this.dryMode}, pause: ${this.pauseMode}. skipping stake`)
       return
     }
     const tx = await this.bridge.stake(amount)
@@ -231,8 +233,9 @@ class StakeWatcher extends BaseWatcher {
       )
     }
     this.logger.debug(`attempting to unstake ${parsedAmount} tokens`)
-    if (this.dryMode) {
-      this.logger.warn('dry mode: skipping unstake transaction')
+    await this.handleStateSwitch()
+    if (this.isDryOrPauseMode) {
+      this.logger.warn(`dry: ${this.dryMode}, pause: ${this.pauseMode}. skipping unstake`)
       return
     }
     const tx = await this.bridge.unstake(amount)
